Spec 002 tache 10 : la panique efface pour de bon
R5.3. Trois fenetres au sous-menu de la barre d'etat, et le choix EST la
confirmation : l'exigence dit « sans justification demandee », et une boite de
dialogue de plus ferait de l'urgence une formalite.

Aucune decoupe. Un episode d'une heure qui touche la fenetre d'une seconde part
en entier — l'immutabilite de la spec 001 interdit de le recouper, et un episode
recoupe ne rejouerait rien.

Trois formes d'episode coexistent sur le disque et chacune se date autrement :
l'episode assemble par ses bornes, l'orphelin par son marqueur, celui en
quarantaine par le ULID de son identifiant. Sans ce dernier, la panique aurait
laisse derriere elle exactement ce que l'operateur veut voir partir.

Un dossier qu'on ne sait pas dater n'est ni efface ni tu : le supprimer
detruirait ce que l'operateur n'a pas demande de detruire, le taire lui
laisserait croire qu'il est parti.

L'attribut lecture seule est leve avant suppression — sinon la panique echouerait
precisement sur ce qu'elle est censee effacer, et le dirait apres coup a
quelqu'un qui croyait le probleme regle. Les empreintes de copie gardees en
memoire partent aussi : sans ca, un collage s'apparierait a une copie que
l'operateur vient de faire effacer.

263 tests Rust. `pnpm verify` vert, clippy strict propre.

+++ apps/desktop/src-tauri/src/uia.rs
@@ -442,7 +442,7 @@ pub fn surfaces_visibles() -> Vec<String> {
 /// bon sens de l'erreur.
 #[cfg(not(test))]
 fn surface_de(el: &uiautomation::UIElement) -> Option<String> {
-    nom_executable(el.get_process_id().ok()? as u32)
+    nom_executable(el.get_process_id().ok()?)
 }
 
 /// Le nom de fichier de l'exécutable d'un processus, en minuscules.
